refactor: buildMenuの純ロジックをlogic/へ分離し10ファイル規約へ戻す
