+++ moorestech_web/webui/src/features/buildMenu/BuildMenuCategoryGrid.tsx
@@ -1,6 +1,6 @@
 import { FadeRule, SlotGrid } from "@/shared/ui";
 import { buildMenuCategoryNameKey, buildMenuSubCategoryNameKey, useI18n } from "@/shared/i18n";
-import type { BuildMenuDisplayEntry, BuildMenuSection } from "./buildMenuGrouping";
+import type { BuildMenuDisplayEntry, BuildMenuSection } from "./logic/buildMenuGrouping";
 import { BuildMenuSlot } from "./BuildMenuSlot";
 import styles from "./style.module.css";
